Move RestApp into package schemaregistry
